pyproject.toml
[project]
name = "to-skip-or-not"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11.9"
dependencies = [
    "matplotlib>=3.10.3",
    "monai~=1.3.1",
    "nibabel~=5.2.1",
    "numpy~=1.26.4",
    "pandas~=2.2.2",
    "pillow~=10.3.0",
    "python-dotenv~=1.0.1",
    "pytorch-lightning~=2.3.0",
    "scikit-image~=0.22.0",
    "scipy~=1.13.1",
    "seaborn>=0.13.2",
    "tensorboard~=2.17.0",
    "torch~=2.3.1",
    "torchvision~=0.18.1",
    "wandb~=0.17.1",
]
